+++ kylie/__init__.py
@@ -2,7 +2,13 @@
 
 """kylie - A module for mapping between JSON and Python classes."""
 
-__all__ = ('Attribute', 'Relation', 'Model', 'AttributeSwitcher')
+__all__ = (
+    'Attribute',
+    'DeserializationError',
+    'MappedModelChoice',
+    'Model',
+    'Relation',
+)
 
 __author__ = 'Mark Smith'
 __email__ = 'mark.smith@practicalpoetry.co.uk'
@@ -10,8 +16,8 @@ __version__ = '0.2.0'
 
 from .kylie import (
     Attribute,
-    AttributeSwitcher,
     DeserializationError,
+    MappedModelChoice,
     Model,
     Relation,
 )