+++ HISTORY.rst
@@ -7,9 +7,13 @@ History
 In Development
 --------------
 
+* MappedModelChoice
 * Change to internal Attribute interface (will lead to minor version bump)
 
   ``_apply_model`` has been replaced with a direct set of ``attr_name``.
+
+* Minor code quality improvements.
+* Documentation improvements.
 
 
 0.2.0 (2015-04-22)